add some sugar to midi status byte handling
Replaced the clunky jill::midi struct with a proper class wrapping an enum of
the most useful midi status bytes that encapsulates the logic for splitting out
the channel (in the lower nib) from the status (in the upper nib) for values
between 0x00-0xe0 (and not for the 0xf0 values, which don't have a channel).
Patched all the compilation units that use midi except jclicker, because
this needs to be fixed up for issue #23.

Another change is that jstim and jstimserver no longer have an option
to set the midi channel, because we're now going to specify that
0x00 and 0x10 are specifically for stimulus onsets and offsets, and 0x01 and
0x11 are for pre- and post-trigger events.